casacore_tables::taql::error

Enum TaqlError

pub enum TaqlError {
    LexError {
        pos: SourcePos,
        message: String,
    },
    ParseError {
        pos: SourcePos,
        message: String,
    },
    UnexpectedEnd {
        message: String,
    },
    TypeError {
        message: String,
    },
    ColumnNotFound {
        name: String,
    },
    UnknownFunction {
        name: String,
    },
    ArgumentCount {
        name: String,
        expected: String,
        got: usize,
    },
    Unsupported {
        message: String,
    },
    Table(String),
    DivisionByZero,
    InsertColumnMismatch {
        columns: usize,
        values: usize,
    },
    JoinTableOpen {
        name: String,
        message: String,
    },
}

All errors produced by the TaQL subsystem.

C++ casacore throws TableError or TableGramError for query failures. This enum provides structured variants with source-position context.

Variants

LexError

The lexer encountered an invalid token.

Fields

pos: SourcePos

message: String

ParseError

The parser encountered unexpected input.

Fields

pos: SourcePos

message: String

UnexpectedEnd

The parser reached end of input unexpectedly.

Fields

message: String

TypeError

An expression evaluated to an unexpected type.

Fields

message: String

ColumnNotFound

A column referenced in a query does not exist.

Fields

name: String

UnknownFunction

A function referenced in a query does not exist.

Fields

name: String

ArgumentCount

A function was called with the wrong number of arguments.

Fields

name: String

expected: String

got: usize

Unsupported

A feature is parsed but not yet implemented.

Fields

message: String

Table(String)

An error from the underlying table operations.

DivisionByZero

Division by zero in expression evaluation.

InsertColumnMismatch

INSERT column count does not match value count.

Fields

columns: usize

values: usize

JoinTableOpen

A JOIN table could not be opened.

Fields

name: String

message: String

Implementations

impl TaqlError

pub fn parse(pos: SourcePos, msg: impl Into<String>) -> Self

Convenience constructor for a parse error at a position.

pub fn unexpected_end(msg: impl Into<String>) -> Self

Convenience constructor for an unexpected-end error.